build.gradle: remove usage of GRADLE_VLC_SRC_DIRS
It was set by vlc-android, based on VLC_OUT_PATH.

VLC_OUT_PATH is set to <vlc>/build-android-<triplet>/ndk and only contains temporary files:
- libvlcjni-modules.c
- libvlcjni-symbols.c
They are not used in the java build.

There is no VLC_OUT_PATH/libs folder after libvlcjni is built.
libvlcjni.so is in <libvlcjni>/jni/libs which we already handled.