buildsystem: ensure APP_STL is used when building libvlc.so
It's only used if any of the sources to compile are C++.
This is not the case as we only build libvlcjni-modules.c and libvlcjni-symbols.c.

It links properly because some pkg-config files are forcing -lc++ when linking.

We can give a hint to ndk-build that the code contains C++ code.